Web UI Conflict between auction page and vote page is solved
